Add transition between animations
Add transition state betwen animations to go from run to idle.

Add code to flip player when running the opposite direction that the sprite image is facing.

This causes issues with the highlighing and will need to be adjusted to a new sprite depending on which way the player is facing.